new sorted-list style contexts
contexts were previously represented as finite partial functions. This
had several benefits, including making it easy to prove contraction and
exchange. However, equality over functions isn't decidable, and
ultimately we'll need decidability of context equality in order to prove
progress.

So this commit changes the approach of contexts from being
finite partial functions to being a sorted list data structure. For any
type A, the list is comprised of (Nat, A) pairs, but the Nat is not the
absolute index of the value, but rather a relative offset from the
previous item in the list. Actually, it's one less than the difference
from the previous item in the list. So the mapping {0 : a, 1 : b, 2 : c}
is represented as [(0 , a), (0 , b), (0 , c)] and the mapping
{1 : a, 3 : b, 6 : c, 10 : d) is represented as
[(1 , a), (1 , b), (2 , c), (3 , d)].

This approach has several advantages. Chiefly, equality is
decidable since it's just a simple list. Also, reflexive equality by way
of _==_ is equivalent to semantic equality, i.e. that the two data
structures represent the same mapping. This is because each unique data
structure represents a unique mapping, due to the fact that an "index"
of 0 actually corresponds to a relative offset of 1. These benefits
would be missing if the data structure had to store proof terms.
Also, we still get exchange and contraction "for free", as was the case
with the finite partial functions.

The only downside I'm aware of is that sublists have completely
different meaning from the superlist - i.e. 1 is in the domain of
[(0 , a), (0 , b)] but it's not in the domain of that list's tail,
[(0 , b)] - also, 0 _is_ in the domain of the tail, even tho 0 was
presumably popped off. This shows that these lists should only ever be
used via the proper mechanisms - _,,_ for insertion/extension and
various theorems that work on C ,, (x , a) (_not_ (x , a) :: C) for
destruction. It would probably be good to use agda's `abstract` feature
to expose only these mechanisms publicly, keeping the list data
structure encapsulated. Also, deletion of items, union, and _##_ are not
currently supported, but they're not currently necessary, either.