redefine IRTpl to correspond closely with how it's written in the paper
The problem is that even defining it in the closest possible way, to the
way it's expressed in the paper, it's still super-verbose, hard to read,
and hard to reason about. In fact, it has become larger (tho perhaps a
bit conceptually simpler or easier to follow).

I don't really see a way out of this - it does seem that I'll have to
change all these definitions to be essentially equivalent to the rules
in the paper, but doing so won't make them comprehensible. Although some
issues are on account of the difficult way in which I've defined List
indexing, even with a form of indexing that's maximally clean and
convenient, there will still be challenges with stuff like asserting
that the lists are all of the right size - generally speaking I think
the chief difficulty is that the sort of notation in the paper is
inherently difficult to formalize, esp. while trying to keep the
formalization as close to the paper as possible. To the extent that's
the case, this formalization (or maybe even the previous one?) may be
about as good as we can get.