Change list indexing from accepting a bounds-check proof to being partial
to get the ith value from l, you used to say
`l [[ i given proof-that-i<length-of-l ]]` . This caused a lot of
annoyances, particularly having to prove proof-irrelevance for the
bounds-check proof term, and then use that proof-irrelevance everywhere.
It seemed that there were several more issues being caused, besides just
that one, so I changed list indexing to not accept any proof term, but
instead to return a Maybe (None if the index is out of bounds). This
didn't turn out to make that much of a difference - some things got
better, some got worse - the judgment definitions themselves are about
as verbose and difficult to read as they were before. At any rate, this
approach is probably more conventional and familiar, so we'll stick with
it for now, and we can revert it later if we decide the old approach was
better for some reason.