Reprove or comment out all theorems in light of recent changes to core
progress is proven. it was pretty ridiculous. do not look at the proof
terms.
Some other theorems were commented out - they will be revived in not too
long, but they're either not critical, no longer applicable/relevant, or
they are still pretty important but not enough to prioritize over the
soundness theorems.